FSM job to remove process var
